+++ src/Exceptions/BadMethodCallException.php
@@ -1,0 +1,13 @@
+<?php
+namespace JRBarnard\Recurrence\Exceptions;
+
+use \BadMethodCallException as OriginalBadMethodCallException;
+
+/**
+ * Class BadMethodCallException
+ * @package JRBarnard\Recurrence\Exceptions
+ */
+class BadMethodCallException extends OriginalBadMethodCallException
+{
+    //
+}