+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## v1.1.0
+### Added
+- Weekly interval accepts constructor arguments
+- Weekly interval add magic every{day} method
+- Add BadMethodCallException for use by __call
+- Weekly interval add magic andEvery{day} method
+- Weekly interval add ofEveryWeek setters
+- Weekly interval add ofEvery{weeks}Week magic setter
+
+### Fixed
+- Errors in readme.md
+
 ## v1.0.0
 ### Added
 - Base src files, composer, git config, readme and changelog files.